feat: create utils folder, move files
