+++ src/js/components/validate.js
